+++ eelbrain/plot/__init__.py
@@ -57,7 +57,7 @@ defaults can be changed with :func:`configure`:
 """
 from ._colors import (colors_for_categorial, colors_for_oneway,
                       colors_for_twoway, ColorBar, ColorGrid, ColorList)
-from ._sensors import SensorMaps, SensorMap2d
+from ._sensors import SensorMaps, SensorMap
 from ._topo import TopoArray, TopoButterfly, Topomap
 from ._uts import UTSStat, UTS, UTSClusters
 from ._utsnd import Array, Butterfly